Implement monitoring by reloading all contents on menu change events.
The appfinder could use a better, modular design. I could hack it to
react on some events without reloading everything but that would result
in messy code or too much work. I guess for now it's fine if all
contents are reloaded/rebuilt whenever something changes in the menu
tree.